"Add CallbackService and use it in GreetingService; update imports and providers"
This diff adds a new file `callback.service.ts` and a new provider `CallbackService` in `bot.module.ts`. It
