+++ src/bot/bot.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common'
 import { BotService } from './bot.service'
-import { GreetingService } from './services/greeting.service'
+import { GreetingService, CallbackService } from './services'
 import { ProjectsModule } from 'src/projects/projects.module'
 
 @Module({
     imports: [ProjectsModule],
-    providers: [BotService, GreetingService],
+    providers: [BotService, GreetingService, CallbackService],
     exports: [],
 })
 export class BotModule {}
