+++ src/bot/services/greeting.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common'
-import { ProjectsService } from 'src/projects/projects.service'
+import { CallbackService } from './callback.service'
 
 @Injectable()
 export class GreetingService {
-    constructor(private readonly projectService: ProjectsService) {}
+    constructor(private readonly projectService: CallbackService) {}
     async greeting(bot, chatId, msg) {
         await bot.sendMessage(
             chatId,
@@ -14,8 +14,6 @@ export class GreetingService {
                         [
                             {
                                 text: 'Edit Text',
-                                // we shall check for this value when we listen
-                                // for "callback_query"
                                 callback_data: 'edit',
                             },
                         ],
@@ -24,22 +22,7 @@ export class GreetingService {
             }
         )
         await bot.on('callback_query', async (callbackQuery) => {
-            const data = await this.projectService.findAll()
-            console.log(data)
-
-            const action = callbackQuery.data
-            const msg = callbackQuery.message
-            const opts = {
-                chat_id: msg.chat.id,
-                message_id: msg.message_id,
-            }
-            let text
-
-            if (action === 'edit') {
-                text = JSON.stringify(data)
-            }
-
-            bot.editMessageText(text, opts)
+            await this.projectService.callback(bot, callbackQuery)
         })
     }
 }
