+++ src/bot/services/index.ts
@@ -1,1 +1,2 @@
 export * from './greeting.service'
+export * from './callback.service'
